requirements.txt
fair-esm>=2.0.0
torch>=2.0.0
biopython>=1.81